LIVE-18119: move getTokensWithFunds to  to also capture in LLD's analytics
